Split apart Capabilities.cs
No code changes, just moving each type to its own file.
