+++ src/ModelContextProtocol/Protocol/Types/ClientCapabilities.cs
@@ -1,0 +1,38 @@
+﻿using ModelContextProtocol.Protocol.Messages;
+using ModelContextProtocol.Server;
+using System.Text.Json.Serialization;
+
+namespace ModelContextProtocol.Protocol.Types;
+
+/// <summary>
+/// Represents the capabilities that a client may support.
+/// <see href="https://github.com/modelcontextprotocol/specification/blob/main/schema/">See the schema for details</see>
+/// </summary>
+public class ClientCapabilities
+{
+    /// <summary>
+    /// Experimental, non-standard capabilities that the client supports.
+    /// </summary>
+    [JsonPropertyName("experimental")]
+    public Dictionary<string, object>? Experimental { get; set; }
+
+    /// <summary>
+    /// Present if the client supports listing roots.
+    /// </summary>
+    [JsonPropertyName("roots")]
+    public RootsCapability? Roots { get; set; }
+
+    /// <summary>
+    /// Present if the client supports sampling from an LLM.
+    /// </summary>
+    [JsonPropertyName("sampling")]
+    public SamplingCapability? Sampling { get; set; }
+
+    /// <summary>Gets or sets notification handlers to register with the client.</summary>
+    /// <remarks>
+    /// When constructed, the client will enumerate these handlers, which may contain multiple handlers per key.
+    /// The client will not re-enumerate the sequence.
+    /// </remarks>
+    [JsonIgnore]
+    public IEnumerable<KeyValuePair<string, Func<JsonRpcNotification, CancellationToken, Task>>>? NotificationHandlers { get; set; }
+}
