+++ src/ModelContextProtocol/Protocol/Types/RootsCapability.cs
@@ -1,0 +1,22 @@
+﻿using ModelContextProtocol.Protocol.Messages;
+using ModelContextProtocol.Server;
+using System.Text.Json.Serialization;
+
+namespace ModelContextProtocol.Protocol.Types;
+
+/// <summary>
+/// Represents the roots capability configuration.
+/// <see href="https://github.com/modelcontextprotocol/specification/blob/main/schema/">See the schema for details</see>
+/// </summary>
+public class RootsCapability
+{
+    /// <summary>
+    /// Whether the client supports notifications for changes to the roots list.
+    /// </summary>
+    [JsonPropertyName("listChanged")]
+    public bool? ListChanged { get; set; }
+
+    /// <summary>Gets or sets the handler for sampling requests.</summary>
+    [JsonIgnore]
+    public Func<ListRootsRequestParams?, CancellationToken, Task<ListRootsResult>>? RootsHandler { get; set; }
+}
